public_status/requirements.txt
Flask
psycopg[binary,pool]
gunicorn
python-dotenv
